==> ML/2_subtitles.ipynb <==
# English Movie Level Prediction

# Подготовка субтитров

---

**Входные данные**

Размеченный список фильмов. Содержит целевой признак – значения уровней CEFR, определенные экспертами.

Файлы с субтитрами к фильмам в формате srt.

---

**Цель**

Добавить в таблицу фильмов поле, содержащее слова из субтитров.

---

**Задачи:**  

- сделать обработку субтитров, удалив из них лишнюю информацию (служебные данные, знаки препинания и т. п.);
- проверить соответствие указанных уровней классификации CEFR;
- разработать новые признаки для проверки их оценить их предполагаемое влияние на качество модели.

<h1>Contents<span class="tocSkip"></span></h1>
<div class="toc"><ul class="toc-item"><li><span><a href="#Initial" data-toc-modified-id="Initial-0"><span class="toc-item-num">0&nbsp;&nbsp;</span>Initial</a></span><ul class="toc-item"><li><span><a href="#Imports" data-toc-modified-id="Imports-0.1"><span class="toc-item-num">0.1&nbsp;&nbsp;</span>Imports</a></span></li><li><span><a href="#Constants" data-toc-modified-id="Constants-0.2"><span class="toc-item-num">0.2&nbsp;&nbsp;</span>Constants</a></span></li><li><span><a href="#Functions" data-toc-modified-id="Functions-0.3"><span class="toc-item-num">0.3&nbsp;&nbsp;</span>Functions</a></span></li><li><span><a href="#Settings" data-toc-modified-id="Settings-0.4"><span class="toc-item-num">0.4&nbsp;&nbsp;</span>Settings</a></span></li></ul></li><li><span><a href="#Подготовка-слов-из-субтитров-к-фильмам" data-toc-modified-id="Подготовка-слов-из-субтитров-к-фильмам-1"><span class="toc-item-num">1&nbsp;&nbsp;</span>Подготовка слов из субтитров к фильмам</a></span><ul class="toc-item"><li><span><a href="#Функция-для-чтения-субтитров-из-файла" data-toc-modified-id="Функция-для-чтения-субтитров-из-файла-1.1"><span class="toc-item-num">1.1&nbsp;&nbsp;</span>Функция для чтения субтитров из файла</a></span></li><li><span><a href="#Функция-для-очистки-текста" data-toc-modified-id="Функция-для-очистки-текста-1.2"><span class="toc-item-num">1.2&nbsp;&nbsp;</span>Функция для очистки текста</a></span></li><li><span><a href="#Функция-для-подготовки-субтитров" data-toc-modified-id="Функция-для-подготовки-субтитров-1.3"><span class="toc-item-num">1.3&nbsp;&nbsp;</span>Функция для подготовки субтитров</a></span></li><li><span><a href="#Чтение-таблицы-с-фильмами" data-toc-modified-id="Чтение-таблицы-с-фильмами-1.4"><span class="toc-item-num">1.4&nbsp;&nbsp;</span>Чтение таблицы с фильмами</a></span></li><li><span><a href="#Добавление-слов-из-субтитров-в-таблицу-фильмов" data-toc-modified-id="Добавление-слов-из-субтитров-в-таблицу-фильмов-1.5"><span class="toc-item-num">1.5&nbsp;&nbsp;</span>Добавление слов из субтитров в таблицу фильмов</a></span></li><li><span><a href="#Удаление-из-таблицы-фильмов-без-субтитров" data-toc-modified-id="Удаление-из-таблицы-фильмов-без-субтитров-1.6"><span class="toc-item-num">1.6&nbsp;&nbsp;</span>Удаление из таблицы фильмов без субтитров</a></span></li><li><span><a href="#Пример-для-ручной-проверки" data-toc-modified-id="Пример-для-ручной-проверки-1.7"><span class="toc-item-num">1.7&nbsp;&nbsp;</span>Пример для ручной проверки</a></span></li></ul></li><li><span><a href="#Корректировка-Level" data-toc-modified-id="Корректировка-Level-2"><span class="toc-item-num">2&nbsp;&nbsp;</span>Корректировка Level</a></span></li><li><span><a href="#Save-dataset-for-modeling" data-toc-modified-id="Save-dataset-for-modeling-3"><span class="toc-item-num">3&nbsp;&nbsp;</span>Save dataset for modeling</a></span></li><li><span><a href="#Эксперименты" data-toc-modified-id="Эксперименты-4"><span class="toc-item-num">4&nbsp;&nbsp;</span>Эксперименты</a></span><ul class="toc-item"><li><span><a href="#New-simple-features" data-toc-modified-id="New-simple-features-4.1"><span class="toc-item-num">4.1&nbsp;&nbsp;</span>New simple features</a></span></li><li><span><a href="#Употребление-слов-в-субтитрах-из-словаря-Oxford" data-toc-modified-id="Употребление-слов-в-субтитрах-из-словаря-Oxford-4.2"><span class="toc-item-num">4.2&nbsp;&nbsp;</span>Употребление слов в субтитрах из словаря Oxford</a></span></li></ul></li><li><span><a href="#Save-dataset-for-modeling" data-toc-modified-id="Save-dataset-for-modeling-5"><span class="toc-item-num">5&nbsp;&nbsp;</span>Save dataset for modeling</a></span></li></ul></div>

---

## Initial

### Imports

import pandas as pd
import os
import warnings

import re
from tqdm import tqdm

### Constants

DATASETS_PATH_LOCAL = 'datasets/'                               # local path to data
DATASETS_PATH_REMOTE = '/datasets/'                             # remote path to data

SUBTITLES_PATH_LOCAL = 'datasets/subtitles/'                    # local path to data (subtitles)
SUBTITLES_PATH_REMOTE = '/datasets/subtitles/'                  # remote path to data (subtitles)

CR = '\n'                                                       # new line char sequence

### Functions

def custom_read_csv(file_name, separator=','):
    """
    reading dataset of .csv format:
       first from local storage;
       if unsuccessful from remote storage.
    """

    path_local = f'{PATH_LOCAL}{file_name}'
    path_remote = f'{PATH_REMOTE}{file_name}'
    
    if os.path.exists(path_local):
        return pd.read_csv(path_local, sep=separator)

    elif os.path.exists(path_remote):
        return pd.read_csv(path_remote, sep=separator)

    else:
        print(f'File "{file_name}" not found at the specified path ')

### Settings

# text styles
class f:
    BOLD = "\033[1m"
    ITALIC = "\033[3m"
    END = "\033[0m"

# Pandas defaults
pd.options.display.max_colwidth = 100
pd.options.display.max_rows = 500
pd.options.display.max_columns = 100
pd.options.display.float_format = '{:.3f}'.format
pd.options.display.colheader_justify = 'left'

# others
warnings.filterwarnings('ignore')

---

## Подготовка слов из субтитров к фильмам

### Функция для чтения субтитров из файла

def read_subtitles(file_name, file_ext='srt', file_encoding='latin-1'):
    """
    Чтение текста из файла file_name.
    Возвращает датафрейм с текстом и статус операции:
        'OK' – файл найден;
        'no subtitles' – файл с субтитрами не найден.
    """

    path_local = f'{SUBTITLES_PATH_LOCAL}{file_name}.{file_ext}'
#     print(f'file {path_local} in process')
    
    if os.path.exists(path_local):
        return pd.read_table(path_local, names=['subs'], encoding=file_encoding), 'OK'

    else:
        print(f'file {f.BOLD}{path_local}{f.END} not found')
        return pd.DataFrame(), 'no subtitles'

### Функция для очистки текста

def clean_text(text):
    '''
    Принятый текст в типа string:
        - переводит в нижний регистр;
        - заменяет на пробел символ перевода строки;
        - заменяет на пробел теги (текст, заключенный в квадратные или фигурные скобки);
        - заменяет на пробел все символы, кроме букв, дефисов и апострофов внутри слов;
        - удаляет все слова из 1-2 букв;
        - заменяет цепочки пробелов на одинарный пробел.
    Возвращает очищенный текст.
    '''
    # приведение значений к нижнему регистру
    text = text.lower()
    
    # удаление перевода строки
    text = text.replace('\n', ' ')
    
    # удаление тегов: частей строк между символами < и >, { и }
    text = re.sub('<.*?>', ' ', text)
    text = re.sub('{.*?}', ' ', text)

    #-------------------------------------------------------------------------------------------------
    # удаление всех символов, кроме букв, пробелов и внутренних дефисов и апострофов
    
    # замена дефисов внутри слов на временный заполнитель (перед удалением небуквенных символов)
    text = re.sub('([a-zA-Z])-([a-zA-Z])', '\\1temporarydefis\\2', text)
    # замена апострофов внутри слов на временный заполнитель (перед удалением небуквенных символов)
    text = re.sub("([a-zA-Z])'([a-zA-Z])", '\\1temporaryapostrof\\2', text)
    # удаление всех символов, кроме букв и пробелов
    text = re.sub('[^a-zA-Z\s]', ' ', text)
    # восстановление внутренних дефисов (замена временного заполнителя на настоящий дефис)
    text = text.replace('temporarydefis', '-')
    # восстановление внутренних апострофов (замена временного заполнителя на настоящий апостроф)
    text = text.replace('temporaryapostrof', "'")
    #-------------------------------------------------------------------------------------------------

    # удаление всех символов, кроме букв и пробелов
#     text = re.sub('[^a-zA-Z\s]', ' ', text)                      # на случай, если дефисы и апострофы тоже нужно удалить
    
    # удаление слов из одной-двух букв (за редким исключением)
#     subtitles = re.sub(r'\s\b\w{1,2}\b\s', ' ', subtitles)       # слово окружено пробелами
    text = re.sub(r'\b\w{1,2}\b', ' ', text)                       # неважно, чем окружено слово
    
    # замена нескольких последовательных пробелов на одинарные
    text = re.sub(' +', ' ', text)
    
    # удаление крайних пробелов (мелочь, но пусть будет)
    text = text.strip()

    return text

### Функция для подготовки субтитров

def prepare_subtitles(movie_name, clean=True):
    '''
    Загружает субтитры для указанного фильма movie_name с помощью функции read_subtitles.
    Выполняет очистку текста от лишней информации (служебные данные, знаки препинания и т.п.) с помощью функции clean_text.
    Возвращает очищенный текст.
    '''
    
    # чтение файла с субтитрами; имя файла совпадает с названием фильма в таблице фильмов
    df, status = read_subtitles(movie_name)
    if status == 'no subtitles':
        return 'no subtitles'
    
    # удаление служебных строк (начинаются с цифры)
    df = df[~df.subs.str.match('^\d')].reset_index(drop=True)
    
    # соединение отдельных субтитров в единый корпус
    subtitles = df.subs.str.cat(sep=' ')
    
    if clean:
        subtitles = clean_text(subtitles)
    
    return subtitles

### Чтение таблицы с фильмами

data = pd.read_excel(DATASETS_PATH_LOCAL + 'movies_labels.xlsx', index_col='id').reset_index(drop=True)

data.sample(3)

### Добавление слов из субтитров в таблицу фильмов

data['Subtitles'] = data.Movie.apply(prepare_subtitles, clean=True)

> Для некоторых фильмов файл с субтитрами не найден. Однозначно определить, что это за фильм, и скачать для него субтитры не удалось.

### Удаление из таблицы фильмов без субтитров

data = data[data.Subtitles != 'no subtitles']

> Удалены фильмы, для которых не найден файл субтитров.

### Пример для ручной проверки

Очищенные субтитры

film_name = 'The.Man.Called.Flintstone.1966.1080p.WEB-DL.DD+2.0.H.264-DAWN'
prepare_subtitles(film_name, clean=True)

Без очистки

prepare_subtitles(film_name, clean=False)

---

## Корректировка Level

Для фильмов с несколькими категориями нужно оставить одну (самую высокую).  
Заменить категории с плюсом на обычные, поскольку плюсовых категорий в CEFR нет.

data.Level = data.Level.apply(lambda Level: Level.replace('+', '')[-2:])

> Плюсы удалены.  
> При наличии нескольких категорий от экспертов для фильма установлена самая старшая категория (последние 2 символа).

---

## Save dataset for modeling

data.to_csv(f'datasets/EDA_movies_subtitles.csv', index=False)

> Предобработку субтитров можно/нужно добавить в пайплайн модели.  
> 
> Если не используются новые сгенерированные признаки, модель работает на новых данных и без предварительной обработки.

data.info()

data.sample(5)

---

## Эксперименты

### New simple features

# # длина текста
# data['text_length'] = data.Subtitles.apply(lambda x: len(x))

# # длина текста без пробелов
# data['text_length_char'] = data.Subtitles.str.replace(' ', '').apply(lambda x: len(x))

# # средняя длина слова
# data['avg_word_length'] = data.text_length_char / (data.text_length - data.text_length_char)

> Эксперименты с моделью показали бесполезность этих новых признаков.

### Употребление слов в субтитрах из словаря Oxford

Будет исследовано распределение слов разных уровней в фильмах разных уровней.

Будет учтено как общее количество употреблений слова (даже если оно повторяется неоднократно), так и распределение уникальныхх слов (каждое слово будет учтено только один раз).

Идеальным резултатом будет выраженная зависимость между словами и фильмамит определенного уровня.

# файл содержит все слова из "American Oxford by CEFR level" с указанным уровнем

data_words = pd.read_csv(f'{DATASETS_PATH_LOCAL}EDA_word_level.csv')

# группировка слов по уровням и преобразование в списки

series_oxford = data_words.groupby('Level').word.agg(list)
series_oxford

# преобразование текста с обработанными субтитрами в список, где каждый элемент – слово

data['subs_words'] = data.Subtitles.apply(lambda x: x.split())

def oxford_count(df, level):
    '''
    df: датафрейм для обработки
    level: уровень, соответствующий "American Oxford by CEFR level"
    
    Добавляет в датафрейм:
        поле вида 'level_sum' с количеством употреблений слов соответствующего уровня;
        поле вида 'level_unq' с количеством уникальных слов соответствующего уровня.
    '''
    
    subs = df['subs_words']
    oxford = series_oxford[level]
    
    df[f'{level}_sum'] = df['subs_words'].apply(lambda subs: sum(i in oxford for i in subs))
    df[f'{level}_unq'] = df['subs_words'].apply(lambda subs: sum(i in subs for i in oxford))
    
    return df

# список всех уровней CEFR, имеющихся в списке фильмов
level_list = sorted(list(data.Level.unique()))


# для каждого возможного уровня добавить в датафрейм:
# количество употреблений слов соответствующего уровня и
# количество уникальных слов соответствующего уровня

for level in tqdm(level_list):
    data = oxford_count(data, level)

# общее количество употреблений слов и уникальных слов всех уровней

data['TOTAL_sum'] = 0
data['TOTAL_unq'] = 0

for level in level_list:
    data['TOTAL_sum'] += data[f'{level}_sum']
    data['TOTAL_unq'] += data[f'{level}_unq']


# подсчет доли разных уровней для употреблений слов и уникальных слов

for level in level_list:
    data[f'{level}_sum_%'] = data[f'{level}_sum'] / data['TOTAL_sum']
    data[f'{level}_unq_%'] = data[f'{level}_unq'] / data['TOTAL_unq']

# удаление ненужных полей

data = data.drop(['subs_words','TOTAL_sum','TOTAL_unq'], axis=1)

for level in level_list:
    data = data.drop(f'{level}_sum', axis=1)     # количество употреблений слов определенного уровня
    data = data.drop(f'{level}_unq', axis=1)     # количество уникальных слов определенного уровня    

# сводная таблица для оценки доли употреблений слов и уникальных слов разных уровней
# в зависимости от уровня фильма, установленного экспертом

data.groupby('Level').mean()

> Пояснение к названиям столбцов:  
> A1_sum_% – доля употребленных слов уровня A1 среди всех слов с известным уровнем.  
> A1_unq_% – доля уникальных слов уровня A1 среди всех слов с известным уровнем.  
> Аналогично и для других уровней.

> Если в полученной таблице изучить столбцы сверху вниз, то видно, что сколько либо заметно отличается только уровень A1, как для уникальных слов из словаря Oxford, так и для общего количества их употреблений.  
> Для остальных уровней доля слов разных уровней примерно одинакова.

> Эксперименты с моделированием подтвердили бесполезность этих новых признаков.

---

## Save dataset for modeling

# data.to_csv(f'datasets/EDA_movies_subtitles.csv', index=False)

> Сохранение датасета с новыми признаками

data.info()

data.sample(5)
