==> ML/3_baseline.ipynb <==
# English Movie Level Prediction

# Baseline-модель

---

**Входные данные**

Таблица с искусственно сгенерированными текстами, содержащими исключительно слова одного уровня.  
Структура таблицы в точности соответствует таблице с данными о фильмах.

---

**Цель**

Создать baseline-модель, не зависящую от оценок экспертов, в качестве нижнего порога качества для последующих моделей.

---

**Задачи:**  

- оценить качество модели по тем же метрикам, которые будут использованы для оценки последующих моделей;
- сделать контрольную проверку на тестовой выборке.

<h1>Contents<span class="tocSkip"></span></h1>
<div class="toc"><ul class="toc-item"><li><span><a href="#Intro" data-toc-modified-id="Intro-0"><span class="toc-item-num">0&nbsp;&nbsp;</span>Intro</a></span></li><li><span><a href="#Initial" data-toc-modified-id="Initial-1"><span class="toc-item-num">1&nbsp;&nbsp;</span>Initial</a></span><ul class="toc-item"><li><span><a href="#Imports" data-toc-modified-id="Imports-1.1"><span class="toc-item-num">1.1&nbsp;&nbsp;</span>Imports</a></span></li><li><span><a href="#Constants" data-toc-modified-id="Constants-1.2"><span class="toc-item-num">1.2&nbsp;&nbsp;</span>Constants</a></span></li><li><span><a href="#Functions" data-toc-modified-id="Functions-1.3"><span class="toc-item-num">1.3&nbsp;&nbsp;</span>Functions</a></span></li><li><span><a href="#Settings" data-toc-modified-id="Settings-1.4"><span class="toc-item-num">1.4&nbsp;&nbsp;</span>Settings</a></span></li></ul></li><li><span><a href="#Read-and-Check-data" data-toc-modified-id="Read-and-Check-data-2"><span class="toc-item-num">2&nbsp;&nbsp;</span>Read and Check data</a></span><ul class="toc-item"><li><span><a href="#Read-data" data-toc-modified-id="Read-data-2.1"><span class="toc-item-num">2.1&nbsp;&nbsp;</span>Read data</a></span></li></ul></li><li><span><a href="#Model" data-toc-modified-id="Model-3"><span class="toc-item-num">3&nbsp;&nbsp;</span>Model</a></span><ul class="toc-item"><li><span><a href="#Подготовка-данных" data-toc-modified-id="Подготовка-данных-3.1"><span class="toc-item-num">3.1&nbsp;&nbsp;</span>Подготовка данных</a></span><ul class="toc-item"><li><span><a href="#Выделение-признаков-и-целевой-переменной" data-toc-modified-id="Выделение-признаков-и-целевой-переменной-3.1.1"><span class="toc-item-num">3.1.1&nbsp;&nbsp;</span>Выделение признаков и целевой переменной</a></span></li><li><span><a href="#Разделение-на-обучающую-и-валидационную-выборки" data-toc-modified-id="Разделение-на-обучающую-и-валидационную-выборки-3.1.2"><span class="toc-item-num">3.1.2&nbsp;&nbsp;</span>Разделение на обучающую и валидационную выборки</a></span></li></ul></li><li><span><a href="#Preprocessing" data-toc-modified-id="Preprocessing-3.2"><span class="toc-item-num">3.2&nbsp;&nbsp;</span>Preprocessing</a></span><ul class="toc-item"><li><span><a href="#Селекторы-числовых-и-категориальных-признаков" data-toc-modified-id="Селекторы-числовых-и-категориальных-признаков-3.2.1"><span class="toc-item-num">3.2.1&nbsp;&nbsp;</span>Селекторы числовых и категориальных признаков</a></span></li><li><span><a href="#Предбработка-числовых-признаков" data-toc-modified-id="Предбработка-числовых-признаков-3.2.2"><span class="toc-item-num">3.2.2&nbsp;&nbsp;</span>Предбработка числовых признаков</a></span></li><li><span><a href="#Предбработка-категориальных-признаков" data-toc-modified-id="Предбработка-категориальных-признаков-3.2.3"><span class="toc-item-num">3.2.3&nbsp;&nbsp;</span>Предбработка категориальных признаков</a></span></li><li><span><a href="#Объединение-предобработки-числовых-и-категориальных-признаков" data-toc-modified-id="Объединение-предобработки-числовых-и-категориальных-признаков-3.2.4"><span class="toc-item-num">3.2.4&nbsp;&nbsp;</span>Объединение предобработки числовых и категориальных признаков</a></span></li><li><span><a href="#Pipelines'-table" data-toc-modified-id="Pipelines'-table-3.2.5"><span class="toc-item-num">3.2.5&nbsp;&nbsp;</span>Pipelines' table</a></span></li></ul></li><li><span><a href="#Подбор-гиперпараметров" data-toc-modified-id="Подбор-гиперпараметров-3.3"><span class="toc-item-num">3.3&nbsp;&nbsp;</span>Подбор гиперпараметров</a></span><ul class="toc-item"><li><span><a href="#Objective-functions-для-Optuna" data-toc-modified-id="Objective-functions-для-Optuna-3.3.1"><span class="toc-item-num">3.3.1&nbsp;&nbsp;</span>Objective functions для Optuna</a></span></li><li><span><a href="#Optuna-call" data-toc-modified-id="Optuna-call-3.3.2"><span class="toc-item-num">3.3.2&nbsp;&nbsp;</span>Optuna call</a></span></li></ul></li></ul></li><li><span><a href="#Анализ-моделей" data-toc-modified-id="Анализ-моделей-4"><span class="toc-item-num">4&nbsp;&nbsp;</span>Анализ моделей</a></span><ul class="toc-item"><li><span><a href="#Чтение-тестовых-данных-(реальных-фильмов-с-субтитрами)" data-toc-modified-id="Чтение-тестовых-данных-(реальных-фильмов-с-субтитрами)-4.1"><span class="toc-item-num">4.1&nbsp;&nbsp;</span>Чтение тестовых данных (реальных фильмов с субтитрами)</a></span></li><li><span><a href="#Additional-metrics" data-toc-modified-id="Additional-metrics-4.2"><span class="toc-item-num">4.2&nbsp;&nbsp;</span>Additional metrics</a></span></li></ul></li></ul></div>

---

## Intro

**Some explanations**

Permanent data tables named like: **data**.  
Temporary data tables named like: **df**.  

The code of the cells are as independent as possible from each other in order to freely manipulate the cells.

Intermediate conclusions are highlighted as follows:

> Intermediate conclusion.

---

## Initial

### Imports

Почистить от лишних импортов !!!

import pandas as pd
import numpy as np

import os
import json
import warnings
import joblib
from datetime import date, time, datetime
from time import time

import matplotlib.pyplot as plt
import seaborn as sns
from matplotlib.colors import LinearSegmentedColormap
import plotly.graph_objects as go
import plotly.io as pio

from sklearn.base import BaseEstimator, TransformerMixin
from sklearn.feature_extraction.text import TfidfVectorizer

from sklearn.compose import ColumnTransformer, make_column_transformer, make_column_selector
from sklearn.preprocessing import StandardScaler, OrdinalEncoder, OneHotEncoder, SplineTransformer, FunctionTransformer
from sklearn.experimental import enable_iterative_imputer
from sklearn.impute import IterativeImputer

from sklearn.model_selection import train_test_split, cross_val_score, StratifiedKFold
from sklearn.pipeline import Pipeline, make_pipeline, FeatureUnion

from nltk.corpus import stopwords

from sklearn.dummy import DummyClassifier
from sklearn.linear_model import RidgeClassifier, SGDClassifier
from sklearn.svm import LinearSVC

from sklearn.metrics import f1_score

import optuna
from optuna.distributions import FloatDistribution, IntDistribution, CategoricalDistribution

### Constants

PATH_LOCAL = 'datasets/'                               # local path to data
PATH_REMOTE = '/datasets/'                             # remote path to data

CR = '\n'                                              # new line

TARGET = 'Level'                                       # target name
SCORING = 'f1_weighted'                                # target metric
VALID_FRAC = 0.0                                       # delayed sampling fraction
N_CV = 5                                               # number of folds during cross-validation

N_TRIALS = 20                                          # max of tries when Optuna optimization run
TIMEOUT = 1000                                         # max time when Optuna optimization run
RANDOM_STATE = RANDOM_SEED = RS = 66                   # random_state

ESTIMATOR_LIST = [                                     # estimators list (comment/uncomment estimator to skip/use)
#                   'DummyClassifier',
                  'RidgeClassifier',
                  'SGDClassifier',
                  'LinearSVC',
                 ]

### Functions

def custom_read_csv(file_name, separator=','):
    """
    reading dataset of .csv format:
       first from local storage;
       if unsuccessful from remote storage.
    """

    path_local = f'{PATH_LOCAL}{file_name}'
    path_remote = f'{PATH_REMOTE}{file_name}'
    
    if os.path.exists(path_local):
        return pd.read_csv(path_local, sep=separator)

    elif os.path.exists(path_remote):
        return pd.read_csv(path_remote, sep=separator)

    else:
        print(f'File "{file_name}" not found at the specified path ')

def var_name(var):
    """
    var name determination
    """
    return [name for name in globals() if globals()[name] is var][0]

def plot_Optuna(study, plot_kind='plot_slice', model_name=''):
    '''
    Дополнительная настройка оригинальных графиков Optuna.
    Например, на графике `plot_slice` изначально цвет точек зависел от номера итерации.
    Теперь они все одинакового цвета и полупрозрачные, лучше видны скопления точек.
    
    study: обученный объект класса OptunaSearchCV
    plot_kind: тип графика Optuna
    model_name: название модели
    '''
    
    if plot_kind == 'plot_slice':
        fig = optuna.visualization.plot_slice(study)
        fig.update_traces(
                          marker_color='Darkgrey',
                          marker_size=3,
                          marker_opacity=0.2,
                          marker_line_width=1,
                          marker_line_color='Black',
                         )
    
    elif plot_kind == 'plot_param_importances':
        fig = optuna.visualization.plot_param_importances(study)
        
    elif plot_kind == 'plot_optimization_history':
        fig = optuna.visualization.plot_optimization_history(study)
        fig.update_traces(
                          marker_size=5,
                          marker_opacity=0.3,
                          marker_line_width=1,
                          marker_line_color='Black',
                         )

    fig.update_layout(
                      title_text=model_name,
                      title_x=0,
                      font_size=10,
                     )    
    fig.show()

def add_model_metrics(models, X_train, Y_train, X_valid, Y_valid, cv=N_CV, scoring_list=['f1'], verbose=True):
    '''
    Принимает:
        датафрейм со списком моделей и их характеристиками;
        два датасета (features and target) – обучающую и валидационную выборки;
        параметр cv для cross_val_score;
        список метрик
        
    Для каждой модели в датафрейме добавляет указанные метрики для обоих датасетов.
    '''

    def cv_score(model, X, Y, scoring, cv):
        invert_koeff = -1 if scoring.split('_')[0] == 'neg' else 1   # инвертирование метрик с приставкой "neg_"
        return invert_koeff * cross_val_score(model, X, Y, scoring=scoring, cv=cv, n_jobs=-1).mean()
    
    if verbose: print(f'Performing cross_val_score with cv={N_CV} and scoring_list={scoring_list}:')
    
    for scoring in scoring_list:
    
        # результаты моделей на обучающей выборке (усреднение на кроссвалидации)
        if verbose: print(f'{scoring}_train...', end=' ')
        models[f'{scoring}_train'] = models.model.apply(cv_score, args=(X_train, Y_train, scoring, cv))
        if verbose: print('done')

        # результаты моделей на тестовой выборке (усреднение на кроссвалидации)
        if verbose: print(f'{scoring}_test...', end=' ')
        models[f'{scoring}_test'] = models.model.apply(cv_score , args=(X_valid, Y_valid, scoring, cv))
        if verbose: print('done')
    
    # оптимальные гиперпараметры
    models['best_params'] = models.study.apply(lambda model: model.best_params)
    
    return models

def extract_final_features(pipeline_model):
    '''
    Принимает пайплайн.
    Возвращает список признаков, на которых обучается финальный estimator пайплайна.
    '''
    feature_list = []
    
    for feature in pipeline_model.steps[-2][1].get_feature_names_out():
        feature_list.append(feature.split('__')[1])

    return feature_list

### Settings

# text styles
class f:
    BOLD = "\033[1m"
    ITALIC = "\033[3m"
    END = "\033[0m"

# defaults for charts

# Matplotlib, Seaborn
PLOT_DPI = 150  # dpi for charts rendering 
sns.set_style('whitegrid', {'axes.facecolor': '0.98', 'grid.color': '0.9', 'axes.edgecolor': '1.0'})
plt.rc(
       'axes',
       labelweight='bold',
       titlesize=16,
       titlepad=10,
      )

# Plotly Graph_Objects
pio.templates['my_theme'] = go.layout.Template(
                                               layout_autosize=True,
                                               # width=900,
                                               layout_height=200,
                                               layout_legend_orientation="h",
                                               layout_margin=dict(t=40, b=40),         # (l=0, r=0, b=0, t=0, pad=0)
                                               layout_template='seaborn',
                                              )
pio.templates.default = 'my_theme'

# colors, color schemes
CMAP_SYMMETRIC = LinearSegmentedColormap.from_list('', ['steelblue', 'aliceblue', 'steelblue'])

# Pandas defaults
pd.options.display.max_colwidth = 100
pd.options.display.max_rows = 500
pd.options.display.max_columns = 100
pd.options.display.float_format = '{:.3f}'.format
pd.options.display.colheader_justify = 'left'

# оформление Optuna
optuna.logging.set_verbosity(optuna.logging.WARNING)  # отключение вывода логов при работе optuna

# others
warnings.filterwarnings('ignore')

---

## Read and Check data

### Read data

data = custom_read_csv('EDA_artficial_subs.csv')

data.sample()

data.info()

---

## Model

### Подготовка данных

#### Выделение признаков и целевой переменной

X = data.drop([TARGET, 'Movie'], axis=1)
Y = data[TARGET]

X.shape, Y.shape

#### Разделение на обучающую и валидационную выборки

Валидационная выборка – часть, отрезанная от train, для локального тестирования модели.

*Примечание: если VALID_FRAC <= 0 для подбора гиперпараметров используется вся обучающая выборка (X, Y). Валидационная выборка в этом случае также равна (X, Y). Создание валидационной выборки в этом случае — лишь формальность (заглушка), и результаты валидации не следует интерпретировать никак.*

if VALID_FRAC > 0 :
    X_train, X_valid, Y_train, Y_valid = train_test_split(X, Y, test_size=VALID_FRAC, stratify=Y, random_state=RS)
else:
    X_train, X_valid, Y_train, Y_valid = X, X, Y, Y

X_train.shape, Y_train.shape, X_valid.shape, Y_valid.shape

### Preprocessing

class TextSelector(BaseEstimator, TransformerMixin):
    '''
    Позволяет выбрать указанный признак для последующей обработки
    '''
    def __init__(self, field):
        self.field = field
    def fit(self, X, Y=None):
        return self
    def transform(self, X):
        return X[self.field]

#### Селекторы числовых и категориальных признаков

num_selector = make_column_selector(dtype_include=np.number)
cat_selector = make_column_selector(dtype_exclude=np.number)

#### Предбработка числовых признаков

num_preprocessor = make_pipeline(
                                 IterativeImputer(initial_strategy='mean', random_state=RS),
                                 StandardScaler(),
                                )

num_transformer = make_column_transformer(
                                          (num_preprocessor, num_selector),
                                          remainder='drop'
                                         )

#### Предбработка категориальных признаков

Категориальный признак только один – содержащий субтитры. По крайней мере, пока.

stopwords_english = list(set(stopwords.words('english')))

text_transformer = Pipeline([
                              ('TEXT_SELECT', TextSelector('Subtitles')),
                              ('TFIDF', TfidfVectorizer(
                                                        decode_error='ignore',
                                                        stop_words=stopwords_english,
                                                        token_pattern=r'(?u)\b[a-z]{3,}\b',      # токены из 3-х и более букв
                                                        smooth_idf=True)
                                                       ),
                             ])

#### Объединение предобработки числовых и категориальных признаков

features_union = FeatureUnion([
                               ('TEXT', text_transformer),
                               ('NUM', num_transformer),
                              ])

#### Pipelines' table

pipelines = [
             Pipeline([
                       ('FU', features_union),
                       ('DC', DummyClassifier())
                      ]),
    
             Pipeline([
                       ('FU', features_union),
                       ('RC', RidgeClassifier(random_state=RS))
                      ]),

             Pipeline([
                       ('FU', features_union),
                       ('SGDC', SGDClassifier(random_state=RS))
                      ]),

             Pipeline([
                       ('FU', features_union),
                       ('LSVC', LinearSVC(random_state=RS))
                      ]),
    
            ]

names = ['DummyClassifier', 'RidgeClassifier','SGDClassifier', 'LinearSVC']

short_names = ['DC', 'RC', 'SGDC', 'LSVC']

models = pd.DataFrame(
                      data={'name': names,
                            'short_name': short_names,
                            'model': pipelines,
                           },
                     )
models

В таблице моделей необходимо оставить только те, что есть в списке ESTIMATOR_LIST (то есть удалить лишние).

for item in range(models.shape[0]):
    if models.loc[item,'name'] not in ESTIMATOR_LIST:
        models = models.drop(item, axis=0)
        
models = models.reset_index(drop=True)

models

### Подбор гиперпараметров

#### Objective functions для Optuna

def objective_DC(trial):

    params = {
              'DC__strategy': trial.suggest_categorical('DC__strategy', ['most_frequent','prior','stratified','uniform']),
             }

    model.set_params(**params)
    cv_SKF = StratifiedKFold(n_splits=N_CV)
    
    return cross_val_score(model, X_train, Y_train, scoring=SCORING, cv=cv_SKF, n_jobs=-1).mean()

def objective_RC(trial):

    params = {
              'RC__alpha': trial.suggest_float('RC__alpha', 0.1, 10.0, log=True),
              'RC__class_weight': trial.suggest_categorical('RC__class_weight', [None,'balanced']),
              'RC__max_iter': trial.suggest_int('RC__max_iter', 100, 1000, log=True),
             }
    
    model.set_params(**params)
    cv_SKF = StratifiedKFold(n_splits=N_CV)
    
    return cross_val_score(model, X_train, Y_train, scoring=SCORING, cv=cv_SKF, n_jobs=-1).mean()

def objective_SGDC(trial):

    params = {
              'SGDC__loss': trial.suggest_categorical('SGDC__loss', ['hinge','log_loss']),
              'SGDC__alpha': trial.suggest_float('SGDC__alpha', 1e-5, 1e-1, log=True),
              'SGDC__class_weight': trial.suggest_categorical('SGDC__class_weight', [None,'balanced']),
              'SGDC__max_iter': trial.suggest_int('SGDC__max_iter', 100, 1000, log=True),
             }
    
    model.set_params(**params)
    cv_SKF = StratifiedKFold(n_splits=5)
    
    return cross_val_score(model, X_train, Y_train, scoring=SCORING, cv=cv_SKF, n_jobs=-1).mean()

def objective_LSVC(trial):

    params = {
              'LSVC__C': trial.suggest_float('LSVC__C', 0.1, 10),
              'LSVC__class_weight': trial.suggest_categorical('LSVC__class_weight', [None,'balanced']),
              'LSVC__max_iter': trial.suggest_int('LSVC__max_iter', 100, 1000, log=True),
              'LSVC__dual': trial.suggest_categorical('LSVC__dual', [True, False]),
             }
    
    model.set_params(**params)
    cv_SKF = StratifiedKFold(n_splits=5)
    
    return cross_val_score(model, X_train, Y_train, scoring=SCORING, cv=cv_SKF, n_jobs=-1).mean()

#### Optuna call

for item in range(models.shape[0]):
    
    print('—' * 60)
    print(f"{CR}{models.loc[item,'name']} hyperparams tuning...")
    
    model = models.loc[item,'model']

    # создание объекта optuna.study
    study = optuna.create_study(
                                study_name=models.loc[item,'name'],
                                direction="maximize",
                                sampler=optuna.samplers.TPESampler(seed=RS)
                               )

    # оптимизация (подбор гиперпараметров)
    study.optimize(eval(f"objective_{models.loc[item,'short_name']}"),
                   n_trials=N_TRIALS, timeout=TIMEOUT, show_progress_bar=True, n_jobs=-1)

    # извлечение и обучение лучшей модели – здесь можно сделать обучение на полном наборе данных (X,Y)
    model.set_params(**study.best_params).fit(X_train, Y_train)

    # сохранение результатов в таблице моделей
    models.loc[item,'model'] = model
    models.loc[item,'study'] = study
#     models.loc[item,'features'] = ', '.join(extract_final_features(model))   # пока недостаточно универсально
    models.loc[item,'score'] = study.best_value

    print(f'{CR}{f.BOLD}{study.study_name}{f.END}{CR}')
    print(f'Количество попыток: {len(study.trials)}')
    print(f'Лучший результат: {f.BOLD}{study.best_value:0.4f}{f.END}{CR}')
    print('Комбинация гиперпараметров:')
    print(json.dumps(study.best_params, indent=1, sort_keys=True), f'{CR}')

## Анализ моделей

### Чтение тестовых данных (реальных фильмов с субтитрами)

data_test = custom_read_csv('EDA_movies_subtitles.csv')

X_test = data_test.drop([TARGET, 'Movie'], axis=1)
Y_test = data_test[TARGET]

X_test.shape, Y_test.shape

### Additional metrics

В дополнение к основной метрике, полученной во время подбора гиперпараметров, можно сравнить метрики, рассчитанные с помощью кроссвалидации. Можно изменить параметры кроссвалидации.  
Метрики рассчитываются как для обучающей, так и валидационной выборки.

# дополнительные метрики моделей
models = add_model_metrics(models, X_train, Y_train, X_test, Y_test, cv=N_CV,
                           scoring_list=['f1_weighted','f1_macro','f1_micro'])

models.drop(['short_name','model','study','best_params'], axis=1)

> Метрики, полученные для тестовой выборки, можно использовать в качестве baseline.
